Fixing ethernet library on Mega.
Because the Mega uses a different pin for SS of the W5100, the SPI library wasn't setting it as an output.  It was only setting the hardware SS pin of the ATmega1280, which is elsewhere on the board.  Adding initSS() to W5100.init() fixed it.


(Filtered from arduino/Arduino@eb47701c9f86b6c5bc7c2f8dc4b252ef95945af7)