Allow the READ PARAMETER command to return our version number. (significant size impact: 14 bytes!)
Initialized "address" to eliminate compiler warning (4 bytes!)

Add "atmega168" as a more accurate target name than "diecimila"
(keep diecimila as well for backward compatibility)

Reduce the .hex and .lst targets that are stored in source control
to the three basics: atmega8, atmega168, atmega328.  The other
targets remain in the makefile and makeall, but will need to be
built from source if wanted.  Which should be less of a problem
now that the source is buildable without installing crosspack.

(Filtered from arduino/Arduino@7b1ee0f1b0192143fffbbed66dc046b6568f4386)