reset the Leonardo board when leaving the bootloader, before starting the sketch
Done so all IO and other registers are properly reinitialized when a new sketch is uploaded or when an existing sketch is started.  Uses a watchdog timeout with a 15 ms period to accomplish the reset.  Bootloader checks the reason for reset and only enumerates as bootloader and enters the programming loop if reset was NOT caused by WDT.

(Filtered from arduino/Arduino@f39ec3772222b0395a3cd165e1f49b2f86620173)