Adding ability to send repeated starts in Wire library.
This adds an additional (boolean) parameter to both endTransmission() and requestFrom(), which specifies whether or not to send a stop condition after the corresponding transmission.  This defaults to true, as in the previous behavior of the library.

http://code.google.com/p/arduino/issues/detail?id=663

(Filtered from arduino/Arduino@2f29f898de9c8375b37269c26948be59a515e801)