Support TIMER1C
Some devices, such as the atmega2560 or the atmega256rfr2 have a timer1c
output. It seems this output is not connected to anything on the Arduino
Mega, but this allows using it on third party hardware nonetheless.

(Filtered from arduino/Arduino@02135349a91b229da5fb82a538a5017b3904975a)