Add Serial_::readBreak() to process SEND_BREAK requests
This allows detecting when the USB host sends a break request and what
the value of the request was. See the comments in USBAPI.h for details.

This just modifies the avr core, not the sam core.

(Filtered from arduino/Arduino@b58f2394005cd073d582b8f90f8ba222aa95b0ee)